[#147] Refactor: 탈퇴 회원 처리 시 예외처리
