+++ src/main/java/umc/GrowIT/Server/domain/UserChallenge.java
@@ -28,7 +28,7 @@ public class UserChallenge extends BaseEntity {
     @Column(name = "thoughts", length = 100)
     private String thoughts;
 
-    @Column(name = "certification_image", length = 255)
+    @Column(name = "certification_image", length = 1024)
     private String certificationImage;
 
     @Enumerated(EnumType.STRING)
