+++ src/main/java/umc/GrowIT/Server/util/SecurityUtils.java
@@ -1,4 +1,4 @@
-package umc.GrowIT.Server.utils;
+package umc.GrowIT.Server.util;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
